Clean up physics bake
